+++ src/main/java/com/etnetera/tremapp/controller/ProjectGroupController.java
@@ -4,7 +4,6 @@ import javax.servlet.http.HttpServletRequest;
 import javax.validation.Valid;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.access.annotation.Secured;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
@@ -20,7 +19,6 @@ import com.etnetera.tremapp.model.mongodb.project.ProjectGroup;
 import com.etnetera.tremapp.model.mongodb.user.Permission;
 import com.etnetera.tremapp.repository.mongodb.project.ProjectGroupRepository;
 import com.etnetera.tremapp.user.UserManager;
-import com.etnetera.tremapp.user.UserRole;
 import com.github.dandelion.datatables.core.ajax.DataSet;
 import com.github.dandelion.datatables.core.ajax.DatatablesCriterias;
 import com.github.dandelion.datatables.core.ajax.DatatablesResponse;
@@ -60,11 +58,11 @@ public class ProjectGroupController implements MenuActivityController {
 		return "page/projectGroup/projectGroupDetail";
 	}
 
-	@Secured({UserRole.ROLE_ADMIN})
 	@RequestMapping(value = "/project-groups/edit/{projectGroupId}", method = RequestMethod.GET)
 	public String editProjectGroup(@PathVariable String projectGroupId, Model model) {
 		ProjectGroup projectGroup = projectGroupRepository.findOne(projectGroupId);
 		ControllerModel.exists(projectGroup, ProjectGroup.class);
+		projectGroup.checkUserPermission(userManager.requireUser(), Permission.ADMIN);
 		ProjectGroupCommand projectGroupCommand = new ProjectGroupCommand();
 		projectGroupCommand.fromProject(projectGroup);
 		model.addAttribute("projectGroup", projectGroup);
@@ -72,12 +70,12 @@ public class ProjectGroupController implements MenuActivityController {
 		return "page/projectGroup/projectGroupEdit";
 	}
 
-	@Secured({UserRole.ROLE_ADMIN})
 	@RequestMapping(value = "/project-groups/edit/{projectGroupId}", method = RequestMethod.POST)
 	public String editProjectGroup(@Valid ProjectGroupCommand projectGroupCommand,
 			BindingResult bindingResult, @PathVariable String projectGroupId, Model model) {
 		ProjectGroup projectGroup = projectGroupRepository.findOne(projectGroupId);
 		ControllerModel.exists(projectGroup, ProjectGroup.class);
+		projectGroup.checkUserPermission(userManager.requireUser(), Permission.ADMIN);
 		if (bindingResult.hasErrors()) {
 			model.addAttribute("projectGroup", projectGroup);
 			return "page/projectGroup/projectGroupEdit";
@@ -87,7 +85,6 @@ public class ProjectGroupController implements MenuActivityController {
 		return "redirect:/project-groups/detail/" + projectGroup.getId();
 	}
 
-	@Secured({UserRole.ROLE_ADMIN})
 	@RequestMapping(value = "/project-groups/create", method = RequestMethod.GET)
 	public String createProjectGroup(Model model) {
 		ProjectGroupCommand projectGroupCommand = new ProjectGroupCommand();
@@ -95,7 +92,6 @@ public class ProjectGroupController implements MenuActivityController {
 		return "page/projectGroup/projectGroupCreate";
 	}
 
-	@Secured({UserRole.ROLE_ADMIN})
 	@RequestMapping(value = "/project-groups/create", method = RequestMethod.POST)
 	public String createProjectGroup(@Valid ProjectGroupCommand projectGroupCommand, BindingResult bindingResult) {
 		if (bindingResult.hasErrors()) {
@@ -103,24 +99,25 @@ public class ProjectGroupController implements MenuActivityController {
 		}
 		ProjectGroup projectGroup = new ProjectGroup();
 		projectGroupCommand.toProject(projectGroup);
+		projectGroup.getUsers().put(userManager.requireUserId(), Permission.OWNER);
 		projectGroupRepository.save(projectGroup);
 		return "redirect:/project-groups/detail/" + projectGroup.getId();
 	}
 
-	@Secured({UserRole.ROLE_ADMIN})
 	@RequestMapping(value = "/project-groups/delete/{projectGroupId}", method = RequestMethod.GET)
 	public String deleteProjectGroup(@PathVariable String projectGroupId, Model model) {
 		ProjectGroup projectGroup = projectGroupRepository.findOne(projectGroupId);
 		ControllerModel.exists(projectGroup, ProjectGroup.class);
+		projectGroup.checkUserPermission(userManager.requireUser(), Permission.OWNER);
 		model.addAttribute("projectGroup", projectGroup);
 		return "page/projectGroup/projectGroupDelete";
 	}
 
-	@Secured({UserRole.ROLE_ADMIN})
 	@RequestMapping(value = "/project-groups/delete/{projectGroupId}", method = RequestMethod.POST)
 	public String deleteProjectGroup(@PathVariable String projectGroupId) {
 		ProjectGroup projectGroup = projectGroupRepository.findOne(projectGroupId);
 		ControllerModel.exists(projectGroup, ProjectGroup.class);
+		projectGroup.checkUserPermission(userManager.requireUser(), Permission.OWNER);
 		projectGroupRepository.delete(projectGroup);
 		return "redirect:/project-groups";
 	}
